wire: make all nodes reply to user requests
Until now, only the running_node could reply to user request. However,
that's cumbersome and limiting. After this commit, both sync_node and
chain_selector can handle user requests made from the handle.
