+++ crates/floresta-wire/src/p2p_wire/chain_selector.rs
@@ -51,12 +51,14 @@ use std::time::Instant;
 
 use bitcoin::block::Header;
 use bitcoin::consensus::deserialize;
+use bitcoin::p2p::message_blockdata::Inventory;
 use bitcoin::p2p::ServiceFlags;
 use bitcoin::BlockHash;
 use floresta_chain::pruned_utreexo::BlockchainInterface;
 use floresta_chain::pruned_utreexo::UpdatableChainstate;
 use floresta_chain::UtreexoBlock;
 use floresta_common::service_flags;
+use log::debug;
 use log::info;
 use log::warn;
 use rustreexo::accumulator::node_hash::BitcoinNodeHash;
@@ -66,6 +68,7 @@ use tokio::time::timeout;
 
 use super::error::WireError;
 use super::node::PeerStatus;
+use super::node_interface::UserRequest;
 use super::peer::PeerMessages;
 use crate::address_man::AddressState;
 use crate::node::periodic_job;
@@ -76,6 +79,7 @@ use crate::node::NodeRequest;
 use crate::node::UtreexoNode;
 use crate::node_context::NodeContext;
 use crate::node_context::PeerId;
+use crate::node_interface::NodeResponse;
 
 #[derive(Debug, Default, Clone)]
 /// A p2p driver that attempts to connect with multiple peers, ask which chain are them following
@@ -645,6 +649,11 @@ where
                         .insert(InflightRequests::Headers, (new_sync_peer, Instant::now()));
                 }
             }
+
+            if let InflightRequests::UserRequest(req) = request {
+                self.user_requests.send_answer(req, None);
+            }
+
             self.inflight.remove(&request);
         }
 
@@ -689,6 +698,8 @@ where
             {
                 try_and_log!(self.handle_notification(notification).await);
             }
+
+            try_and_log!(self.handle_user_request().await);
 
             periodic_job!(
                 self.maybe_open_connection().await,
@@ -842,6 +853,45 @@ where
             PeerMessages::Addr(addresses) => {
                 let addresses: Vec<_> = addresses.iter().cloned().map(|addr| addr.into()).collect();
                 self.address_man.push_addresses(&addresses);
+            }
+
+            PeerMessages::Block(block) => {
+                if self.check_is_user_block_and_reply(block).await?.is_some() {
+                    log::error!("peer {peer} sent us a block we didn't request");
+                    self.increase_banscore(peer, 5).await?;
+                }
+            }
+
+            PeerMessages::NotFound(inv) => match inv {
+                Inventory::Error => {}
+                Inventory::Block(block)
+                | Inventory::WitnessBlock(block)
+                | Inventory::CompactBlock(block) => {
+                    self.user_requests
+                        .send_answer(UserRequest::Block(block), None);
+                }
+
+                Inventory::WitnessTransaction(tx) | Inventory::Transaction(tx) => {
+                    self.user_requests
+                        .send_answer(UserRequest::MempoolTransaction(tx), None);
+                }
+                _ => {}
+            },
+
+            PeerMessages::Transaction(tx) => {
+                debug!("saw a mempool transaction with txid={}", tx.compute_txid());
+                self.user_requests.send_answer(
+                    UserRequest::MempoolTransaction(tx.compute_txid()),
+                    Some(NodeResponse::MempoolTransaction(tx)),
+                );
+            }
+
+            PeerMessages::UtreexoState(_) => {
+                warn!(
+                    "Utreexo state received from peer {}, but we didn't ask",
+                    peer
+                );
+                self.increase_banscore(peer, 5).await?;
             }
 
             _ => {}
